GenerateHTMLButton.CopyImagesFromAppDataLocalTemp.csを追加 以下のように仕様変更 AppData/Local/Tempの画像をImgFromTempにコピー AppData/Local/Tempの画像をpictにコピー
